requirements.txt
flask>=2.0.1
flask-socketio>=5.1.1
openai==1.3.0
pyyaml==6.0
faster-whisper
torch